Updating fr+de models, training, features, documentation

+++ src/main/java/com/scienceminer/nerd/features/GenericSelectionFeatureVector.java
@@ -270,7 +270,7 @@ public class GenericSelectionFeatureVector {
 			i++;
 		}
 		if (Add_nb_tokens) {
-			result[i] = nb_tokens;
+			result[i] = (double)nb_tokens;
 			i++;
 		}
 		if (Add_relatedness) {
